# k_admin/admin-web/apps/web-antd/src/api/kadmin/generated/ebook_download.ts
// Code generated by kadmin codegen. DO NOT EDIT.
import { request } from '../client';

export interface EbookDownload {
  downloadId: number;
  ebookId: number;
  userId: number;
  createBy: string;
  createTime: string;
  updateBy: string;
  updateTime: string;
  remark: string;
}

export interface EbookDownloadPayload {
  ebookId: number;
  userId: number;
  createBy: string;
  createTime: string;
  updateBy: string;
  updateTime: string;
  remark: string;
}

export interface EbookDownloadFilters {
  page?: number;
  pageSize?: number;
  keyword?: string;
  createBy?: string;
  createTime?: string;
  updateBy?: string;
  updateTime?: string;
  remark?: string;
}

export interface PageResult<T> {
  items: T[];
  page: number;
  pageSize: number;
  total: number;
}

function queryString<T extends object>(filters: T) {
  const params = new URLSearchParams();
  Object.entries(filters).forEach(([key, value]) => {
    if (value !== undefined && value !== '') {
      params.set(key, String(value));
    }
  });
  const query = params.toString();
  return query ? `?${query}` : '';
}

export function getEbookDownloadList(filters: EbookDownloadFilters = {}) {
  return request<PageResult<EbookDownload>>(`/api/ebook-downloads${queryString(filters)}`);
}

export function getEbookDownload(id: number) {
  return request<EbookDownload>(`/api/ebook-downloads/${id}`);
}

export function createEbookDownload(payload: EbookDownloadPayload) {
  return request<EbookDownload>(`/api/ebook-downloads`, {
    method: 'POST',
    body: JSON.stringify(payload),
  });
}

export function updateEbookDownload(id: number, payload: EbookDownloadPayload) {
  return request<EbookDownload>(`/api/ebook-downloads/${id}`, {
    method: 'PUT',
    body: JSON.stringify(payload),
  });
}

export function deleteEbookDownload(id: number) {
  return request<boolean>(`/api/ebook-downloads/${id}`, { method: 'DELETE' });
}