# k_admin/internal/kadmin/generated/ebook_download/routes.go
// Code generated by kadmin codegen. DO NOT EDIT.

package ebook_download

import (
	"errors"
	"fmt"
	"net/http"
	"strconv"
	"strings"

	"github.com/GoAdminGroup/go-admin/internal/kadmin/transport/httpx"
	"github.com/GoAdminGroup/go-admin/modules/db"
	"github.com/gin-gonic/gin"
)

// Dependencies wires the generated module into the KAdmin API group.
type Dependencies struct {
	Connection              db.Connection
	RequireAuth             gin.HandlerFunc
	RequirePermission       func(...string) gin.HandlerFunc
	RegisterAuditResource   func(prefix, resource string, loader func(string) interface{})
	RegisterIdempotentRoute func(prefix string)
}

// Register installs the CRUD routes for table ptmj_ebook_download and keeps the
// module permission and menu rows in sync.
func Register(api *gin.RouterGroup, deps Dependencies) error {
	repo := newRepository(deps.Connection)
	if err := ensurePermissions(deps.Connection); err != nil {
		return fmt.Errorf("sync ebook_download permissions: %w", err)
	}
	if err := ensureMenu(deps.Connection); err != nil {
		return fmt.Errorf("sync ebook_download menu: %w", err)
	}
	deps.RegisterAuditResource("ebook-downloads", "ebook_download", func(resourceID string) interface{} {
		id, _ := strconv.ParseInt(resourceID, 10, 64)
		if id <= 0 {
			return nil
		}
		item, found, err := repo.findByID(id)
		if err != nil || !found {
			return nil
		}
		return item
	})
	deps.RegisterIdempotentRoute("ebook-downloads")
	h := &handler{repo: repo}
	group := api.Group("/ebook-downloads", deps.RequireAuth)
	group.GET("", deps.RequirePermission(ListPermission), h.list)
	group.GET("/:id", deps.RequirePermission(ListPermission), h.get)
	group.POST("", deps.RequirePermission(CreatePermission), h.create)
	group.PUT("/:id", deps.RequirePermission(UpdatePermission), h.update)
	group.DELETE("/:id", deps.RequirePermission(DeletePermission), h.delete)
	return nil
}

type handler struct {
	repo *repository
}

// swaggerListEbookDownload documents GET /ebook-downloads.
// @Summary 查询电子书下载记录列表
// @Tags 生成业务
// @Security BearerAuth
// @Param page query int false "页码"
// @Param pageSize query int false "每页数量"
// @Param keyword query string false "关键词"
// @Success 200 {object} object
// @Failure 401 {object} object
// @Failure 403 {object} object
// @Router /ebook-downloads [get]
func (h *handler) list(c *gin.Context) {
	filter := EbookDownloadFilter{
		Page:       positiveInt(c.Query("page"), 1),
		PageSize:   positiveInt(c.Query("pageSize"), 20),
		Keyword:    truncate(c.Query("keyword"), 100),
		CreateBy:   truncate(c.Query("createBy"), 100),
		CreateTime: truncate(c.Query("createTime"), 100),
		UpdateBy:   truncate(c.Query("updateBy"), 100),
		UpdateTime: truncate(c.Query("updateTime"), 100),
		Remark:     truncate(c.Query("remark"), 100),
	}
	if filter.PageSize > 100 {
		filter.PageSize = 100
	}
	page, err := h.repo.list(filter)
	if err != nil {
		httpx.Fail(c, http.StatusInternalServerError, err.Error())
		return
	}
	httpx.Success(c, page)
}

// swaggerGetEbookDownload documents GET /ebook-downloads/{id}.
// @Summary 获取电子书下载记录详情
// @Tags 生成业务
// @Security BearerAuth
// @Param id path int true "记录 ID"
// @Success 200 {object} object
// @Failure 401 {object} object
// @Failure 403 {object} object
// @Failure 404 {object} object
// @Router /ebook-downloads/{id} [get]
func (h *handler) get(c *gin.Context) {
	id, ok := pathID(c)
	if !ok {
		return
	}
	item, found, err := h.repo.findByID(id)
	if err != nil {
		httpx.Fail(c, http.StatusInternalServerError, err.Error())
		return
	}
	if !found {
		httpx.Fail(c, http.StatusNotFound, errEbookDownloadNotFound.Error())
		return
	}
	httpx.Success(c, item)
}

// swaggerCreateEbookDownload documents POST /ebook-downloads.
// @Summary 新增电子书下载记录
// @Tags 生成业务
// @Security BearerAuth
// @Param payload body EbookDownloadPayload true "电子书下载记录信息"
// @Param Idempotency-Key header string true "幂等键（8-128 位）"
// @Success 200 {object} object
// @Failure 400 {object} object
// @Failure 401 {object} object
// @Failure 403 {object} object
// @Failure 409 {object} object
// @Router /ebook-downloads [post]
func (h *handler) create(c *gin.Context) {
	var payload EbookDownloadPayload
	if err := c.ShouldBindJSON(&payload); err != nil {
		httpx.Fail(c, http.StatusBadRequest, "invalid ebook_download payload")
		return
	}
	if message := validateEbookDownloadPayload(payload); message != "" {
		httpx.Fail(c, http.StatusBadRequest, message)
		return
	}
	item, err := h.repo.create(payload)
	if err != nil {
		respondEbookDownloadError(c, err)
		return
	}
	httpx.Success(c, item)
}

// swaggerUpdateEbookDownload documents PUT /ebook-downloads/{id}.
// @Summary 修改电子书下载记录
// @Tags 生成业务
// @Security BearerAuth
// @Param id path int true "记录 ID"
// @Param payload body EbookDownloadPayload true "电子书下载记录信息"
// @Success 200 {object} object
// @Failure 400 {object} object
// @Failure 401 {object} object
// @Failure 403 {object} object
// @Failure 404 {object} object
// @Router /ebook-downloads/{id} [put]
func (h *handler) update(c *gin.Context) {
	id, ok := pathID(c)
	if !ok {
		return
	}
	var payload EbookDownloadPayload
	if err := c.ShouldBindJSON(&payload); err != nil {
		httpx.Fail(c, http.StatusBadRequest, "invalid ebook_download payload")
		return
	}
	if message := validateEbookDownloadPayload(payload); message != "" {
		httpx.Fail(c, http.StatusBadRequest, message)
		return
	}
	item, err := h.repo.update(id, payload)
	if err != nil {
		respondEbookDownloadError(c, err)
		return
	}
	httpx.Success(c, item)
}

// swaggerDeleteEbookDownload documents DELETE /ebook-downloads/{id}.
// @Summary 删除电子书下载记录
// @Tags 生成业务
// @Security BearerAuth
// @Param id path int true "记录 ID"
// @Success 200 {object} object
// @Failure 401 {object} object
// @Failure 403 {object} object
// @Failure 404 {object} object
// @Router /ebook-downloads/{id} [delete]
func (h *handler) delete(c *gin.Context) {
	id, ok := pathID(c)
	if !ok {
		return
	}
	if err := h.repo.delete(id); err != nil {
		respondEbookDownloadError(c, err)
		return
	}
	httpx.Success(c, true)
}

func validateEbookDownloadPayload(payload EbookDownloadPayload) string {
	return ""
}

func respondEbookDownloadError(c *gin.Context, err error) {
	switch {
	case errors.Is(err, errEbookDownloadNotFound):
		httpx.Fail(c, http.StatusNotFound, err.Error())
	case strings.Contains(err.Error(), "duplicate key"):
		httpx.Fail(c, http.StatusConflict, "记录已存在")
	default:
		httpx.Fail(c, http.StatusInternalServerError, err.Error())
	}
}

// ensurePermissions inserts the four CRUD permission rows when missing.
func ensurePermissions(conn db.Connection) error {
	seeds := []struct{ name, slug, method, path string }{
		{"电子书下载记录查看", ListPermission, "GET", "/api/ebook-downloads*"},
		{"电子书下载记录新增", CreatePermission, "POST", "/api/ebook-downloads*"},
		{"电子书下载记录修改", UpdatePermission, "PUT", "/api/ebook-downloads*"},
		{"电子书下载记录删除", DeletePermission, "DELETE", "/api/ebook-downloads*"},
	}
	for _, seed := range seeds {
		rows, err := conn.Query(`SELECT id FROM goadmin_permissions WHERE slug = ?`, seed.slug)
		if err != nil {
			return err
		}
		if len(rows) > 0 {
			continue
		}
		if _, err := conn.Exec(`INSERT INTO goadmin_permissions (name, slug, http_method, http_path, created_at, updated_at)
			VALUES (?, ?, ?, ?, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)`,
			seed.name, seed.slug, seed.method, seed.path); err != nil {
			return err
		}
	}
	return nil
}

// ensureMenu inserts the module menu under the shared /business directory.
func ensureMenu(conn db.Connection) error {
	parentRows, err := conn.Query(`SELECT id FROM goadmin_menu WHERE uri = ?`, "/business")
	if err != nil || len(parentRows) == 0 {
		if err == nil {
			err = fmt.Errorf("parent menu /business is missing")
		}
		return err
	}
	parentID := toInt64(parentRows[0]["id"])
	rows, err := conn.Query(`SELECT id FROM goadmin_menu WHERE uri = ?`, "/ebook-downloads")
	if err != nil {
		return err
	}
	if len(rows) > 0 {
		return nil
	}
	_, err = conn.Exec(`INSERT INTO goadmin_menu
		(parent_id, type, "order", title, icon, uri, plugin_name, component, created_at, updated_at)
		VALUES (?, 1, 99, ?, 'lucide:package-open', ?, '', ?, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)`,
		parentID, "电子书下载记录", "/ebook-downloads", "/kadmin/generated/ebook_download/EbookDownloadListView")
	return err
}

func pathID(c *gin.Context) (int64, bool) {
	id, err := strconv.ParseInt(c.Param("id"), 10, 64)
	if err != nil || id <= 0 {
		httpx.Fail(c, http.StatusBadRequest, "invalid id")
		return 0, false
	}
	return id, true
}

func positiveInt(value string, fallback int) int {
	parsed, err := strconv.Atoi(strings.TrimSpace(value))
	if err != nil || parsed <= 0 {
		return fallback
	}
	return parsed
}

func truncate(value string, limit int) string {
	value = strings.TrimSpace(value)
	runes := []rune(value)
	if len(runes) > limit {
		return string(runes[:limit])
	}
	return value
}