# k_admin/internal/kadmin/generated/ebook_download/types.go
// Code generated by kadmin codegen. DO NOT EDIT.

package ebook_download

// Permission codes for table ptmj_ebook_download.
const (
	ListPermission   = "system:ebook_download:list"
	CreatePermission = "system:ebook_download:create"
	UpdatePermission = "system:ebook_download:update"
	DeletePermission = "system:ebook_download:delete"
)

// EbookDownload is the model of table ptmj_ebook_download.
type EbookDownload struct {
	DownloadId int64  `json:"downloadId"`
	EbookId    int64  `json:"ebookId"`
	UserId     int64  `json:"userId"`
	CreateBy   string `json:"createBy"`
	CreateTime string `json:"createTime"`
	UpdateBy   string `json:"updateBy"`
	UpdateTime string `json:"updateTime"`
	Remark     string `json:"remark"`
}

// EbookDownloadPayload carries the writable fields of EbookDownload.
type EbookDownloadPayload struct {
	EbookId    int64  `json:"ebookId"`
	UserId     int64  `json:"userId"`
	CreateBy   string `json:"createBy"`
	CreateTime string `json:"createTime"`
	UpdateBy   string `json:"updateBy"`
	UpdateTime string `json:"updateTime"`
	Remark     string `json:"remark"`
}

// EbookDownloadFilter carries the list query parameters.
type EbookDownloadFilter struct {
	Page       int    `json:"page"`
	PageSize   int    `json:"pageSize"`
	Keyword    string `json:"keyword"`
	CreateBy   string `json:"createBy"`
	CreateTime string `json:"createTime"`
	UpdateBy   string `json:"updateBy"`
	UpdateTime string `json:"updateTime"`
	Remark     string `json:"remark"`
}

// EbookDownloadPage is a paginated list result.
type EbookDownloadPage struct {
	Items    []EbookDownload `json:"items"`
	Total    int64           `json:"total"`
	Page     int             `json:"page"`
	PageSize int             `json:"pageSize"`
}